Desync fixes
Write_VDP_Ctrl is used to emulate sending data to the VDP control port.
So literally at the moment the savestate is loaded this code was sending data to VDP port,
thus it eventually in some cases was triggering DMA Update,
or causing other issues.
Also VDP_Status backup for fake VDP render.